requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
pyyaml>=6.0.1
httpx>=0.27.0
openai>=1.30.0
anthropic>=0.49.0
pytest>=8.1.0
pytest-asyncio>=0.23.0
